tests/test_project/requirements.txt
Django>=4.2
djangorestframework
redis
pymongo
# django-smart-ratelimit is installed separately via `pip install -e .` from repo root
requests
